+++ code/motor.h
@@ -4,8 +4,8 @@
 #define PWM1                    (ATOM0_CH4_P02_4)
 #define PWM2                    (ATOM0_CH5_P02_5)
 
-extern int duty1;
-extern int duty2;
+extern int16_t duty1;
+extern int16_t duty2;
 
 void Motor_init(void);
 void Motor_set_duty(int duty1, int duty2);
